requirements.txt
Flask==2.0.1
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==20.1.0 